+++ utils/.internal/ammendParams.js
@@ -1,0 +1,10 @@
+const ammendParams = (params = {}, providerName, providerSettings) => {
+  if (!providerSettings[providerName]) return params;
+
+  return {
+    ...params,
+    ...providerSettings[providerName]
+  };
+};
+
+module.exports = ammendParams
